chore(app): move to SeaORM from Prisma

+++ src/api/expenses/dto/create_expense_db_dto.rs
@@ -1,11 +1,14 @@
+use sea_orm::{DeriveIntoActiveModel};
+use sea_orm::prelude::Decimal;
 use serde::{Deserialize, Serialize};
-
 use crate::api::expenses::types::expense_category::ExpenseCategory;
 
-#[derive(Serialize, Deserialize, Debug)]
+use crate::shared::modules::db::entities::expense::ActiveModel;
+
+#[derive(Serialize, Deserialize, Debug, DeriveIntoActiveModel)]
 pub struct CreateExpenseDbDto {
     pub customer_id: String,
-    pub amount: f64,
+    pub amount: Decimal,
     pub date: chrono::DateTime<chrono::FixedOffset>,
     pub category: ExpenseCategory,
 }
